Support for big PDFs over 2G. (Initial contribution by Welman Jordan)

+++ src/main/java/com/itextpdf/text/pdf/OutputStreamCounter.java
@@ -52,7 +52,7 @@ import java.io.OutputStream;
 public class OutputStreamCounter extends OutputStream {
     
     protected OutputStream out;
-    protected int counter = 0;
+    protected long counter = 0;
     
     /** Creates a new instance of OutputStreamCounter */
     public OutputStreamCounter(OutputStream out) {
@@ -157,7 +157,7 @@ public class OutputStreamCounter extends OutputStream {
         out.write(b, off, len);
     }
     
-    public int getCounter() {
+    public long getCounter() {
         return counter;
     }
